# Extend critical section in commit to fix race condition