Race condition where a dependent commit is persistent before the commit it depends on, which will result in inconsistent state to the external world (or when we recover from log) if a crash happens immediately when the first commit is  written out but not the second one.  (More detail as comment in code)

By extending the critical section we rule this out from happening. This is overly strict and does not work if we go to serializable, but we can cross that bridge when we come to it.

Performance-wise, little to no impact (within margin of error) on low thread numbers because the global transaction table maintenance is the heavy-hitter, but some impact as we scale up (say, 16 threads). Again, can always fix later. Good numbers on incorrect software is useless.

Numbers from dev 4 with write ahead logging on:
- 4 threads
before:
```
tli2@dev4:~/terrier/build$ numactl --cpunodebind=0 --membind=0 release/logging_benchmark
2018-10-15 14:48:48
Running release/logging_benchmark
Run on (40 X 3100 MHz CPU s)
CPU Caches:
  L1 Data 32K (x20)
  L1 Instruction 32K (x20)
  L2 Unified 256K (x20)
  L3 Unified 25600K (x2)
***WARNING*** CPU scaling is enabled, the benchmark real time measurements may be noisy and will incur extra overhead.
---------------------------------------------------------------------------------------------------------
Benchmark                                                                  Time           CPU Iterations
---------------------------------------------------------------------------------------------------------
LoggingBenchmark/TPCCish/min_time:2.000/manual_time                      339 ms       1997 ms          8   287.753k items/s
LoggingBenchmark/SingleStatementInsert/min_time:2.000/manual_time        199 ms       2017 ms         12   491.146k items/s
LoggingBenchmark/SingleStatementUpdate/min_time:2.000/manual_time        187 ms       1949 ms         14   523.426k items/s
LoggingBenchmark/SingleStatementSelect/min_time:2.000/manual_time        168 ms       1956 ms         17   581.084k items/s
```
after:
```
tli2@dev4:~/terrier/build$ numactl --cpunodebind=1 --membind=1 release/logging_benchmark
2018-10-15 14:37:11
Running release/logging_benchmark
Run on (40 X 3100 MHz CPU s)
CPU Caches:
  L1 Data 32K (x20)
  L1 Instruction 32K (x20)
  L2 Unified 256K (x20)
  L3 Unified 25600K (x2)
***WARNING*** CPU scaling is enabled, the benchmark real time measurements may be noisy and will incur extra overhead.
---------------------------------------------------------------------------------------------------------
Benchmark                                                                  Time           CPU Iterations
---------------------------------------------------------------------------------------------------------
LoggingBenchmark/TPCCish/min_time:2.000/manual_time                      343 ms       1972 ms          8   284.505k items/s
LoggingBenchmark/SingleStatementInsert/min_time:2.000/manual_time        204 ms       1966 ms         14    478.54k items/s
LoggingBenchmark/SingleStatementUpdate/min_time:2.000/manual_time        188 ms       1988 ms         15   520.186k items/s
LoggingBenchmark/SingleStatementSelect/min_time:2.000/manual_time        161 ms       1974 ms         17   606.118k items/s
```
- 16 threads
before:
```
tli2@dev4:~/terrier/build$ numactl --cpunodebind=1 --membind=1 release/logging_benchmark
2018-10-15 14:44:51
Running release/logging_benchmark
Run on (40 X 3100 MHz CPU s)
CPU Caches:
  L1 Data 32K (x20)
  L1 Instruction 32K (x20)
  L2 Unified 256K (x20)
  L3 Unified 25600K (x2)
***WARNING*** CPU scaling is enabled, the benchmark real time measurements may be noisy and will incur extra overhead.
---------------------------------------------------------------------------------------------------------
Benchmark                                                                  Time           CPU Iterations
---------------------------------------------------------------------------------------------------------
LoggingBenchmark/TPCCish/min_time:2.000/manual_time                      228 ms       2036 ms         10   428.693k items/s
LoggingBenchmark/SingleStatementInsert/min_time:2.000/manual_time        200 ms       1948 ms         11   488.281k items/s
LoggingBenchmark/SingleStatementUpdate/min_time:2.000/manual_time        177 ms       1967 ms         16   551.341k items/s
LoggingBenchmark/SingleStatementSelect/min_time:2.000/manual_time        176 ms       1964 ms         16   555.062k items/s
```
after:
```
tli2@dev4:~/terrier/build$ numactl --cpunodebind=1 --membind=1 release/logging_benchmark
2018-10-15 14:41:54
Running release/logging_benchmark
Run on (40 X 3100 MHz CPU s)
CPU Caches:
  L1 Data 32K (x20)
  L1 Instruction 32K (x20)
  L2 Unified 256K (x20)
  L3 Unified 25600K (x2)
***WARNING*** CPU scaling is enabled, the benchmark real time measurements may be noisy and will incur extra overhead.
---------------------------------------------------------------------------------------------------------
Benchmark                                                                  Time           CPU Iterations
---------------------------------------------------------------------------------------------------------
LoggingBenchmark/TPCCish/min_time:2.000/manual_time                      230 ms       2011 ms          9   424.387k items/s
LoggingBenchmark/SingleStatementInsert/min_time:2.000/manual_time        200 ms       1978 ms         11   487.616k items/s
LoggingBenchmark/SingleStatementUpdate/min_time:2.000/manual_time        197 ms       2028 ms         12   495.507k items/s
LoggingBenchmark/SingleStatementSelect/min_time:2.000/manual_time        200 ms       1940 ms         13   488.657k items/s
```